Save Name
Name changes on left side, but not on right side when going to a different function and coming back. Also, save needs to keep the text info in the file while just changing the file name (currently makes new file path without moving the contents of the file into it).